+++ src/main/java/com/nautik/api/dto/port/company/CompanyDto.java
@@ -11,10 +11,10 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CompanyDto  {
-    private Long id;
+    private Integer id;
     private String name;
     private String vat;
     private String email;
     private String phone;
-    private Long adminId;
+    private Integer adminId;
 }
